[refactor] 고객 페이지 분리
